fixes #22
