+++ requirements.txt
@@ -1,3 +1,3 @@
-tm-python @ git+https://github.com/cms-l1-globaltrigger/tm-python@0.9.1
+tm-python @ git+https://github.com/cms-l1-globaltrigger/tm-python@0.10.0
 Markdown>=3.1
 PyQt5==5.13.*
