+++ tests/test_menu.py
@@ -19,7 +19,7 @@ import unittest
 import tempfile
 import logging
 
-UTM_VERSION = '0.9.1'
+UTM_VERSION = '0.10.0'
 
 class MenuTests(unittest.TestCase):
 
@@ -105,6 +105,7 @@ class MenuTests(unittest.TestCase):
             "MU10": [],
             "MU80[MU-ISO_Q,MU-QLTY_Q]": ["MU-ISO_Q", "MU-QLTY_Q"],
             "TAU10[TAU-ETA_Q,TAU-PHI_Q]": ["TAU-ETA_Q", "TAU-PHI_Q"],
+            "JET0[JET_DISP_LLP]": ["JET-DISP_LLP"],
         }
         for token, ref in cases.items():
             self.assertEqual(objectCuts(token), ref)
